[svn] Modifs Carcans 2005 Ajout commentaires PAssage numdc en unsigned int pour limiter mémoire

+++ valeur.h
@@ -13,8 +13,10 @@ public:
     ~valeur () {}
 
 public:
-    int   nb;
+    unsigned short   nb;
     float lien;
+    //bool  contig;
+    //int   dist;
 }
 ;
 #endif
